requirements.txt
matplotlib==3.8.3
numpy==2.2.6
opencv_python==4.9.0.80
Pillow==11.2.1
torch==2.6.0
torchvision==0.21.0
